copr: refactor MAX_RAND_VALUE to u32

+++ components/tidb_query/src/util.rs
@@ -5,7 +5,7 @@ use tipb::ColumnInfo;
 
 use crate::codec::datum::Datum;
 
-const MAX_RAND_VALUE: u64 = 0x3FFFFFFF;
+const MAX_RAND_VALUE: u32 = 0x3FFFFFFF;
 
 pub struct MySQLRng {
     seed1: u32,
@@ -20,15 +20,14 @@ impl MySQLRng {
     }
 
     pub fn new_with_seed(seed: i64) -> Self {
-        let seed1 =
-            (seed.wrapping_mul(0x10001).wrapping_add(55555555)) as u32 % MAX_RAND_VALUE as u32;
-        let seed2 = (seed.wrapping_mul(0x10000001)) as u32 % MAX_RAND_VALUE as u32;
+        let seed1 = (seed.wrapping_mul(0x10001).wrapping_add(55555555)) as u32 % MAX_RAND_VALUE;
+        let seed2 = (seed.wrapping_mul(0x10000001)) as u32 % MAX_RAND_VALUE;
         MySQLRng { seed1, seed2 }
     }
 
     pub fn gen(&mut self) -> f64 {
-        self.seed1 = (self.seed1 * 3 + self.seed2) % MAX_RAND_VALUE as u32;
-        self.seed2 = (self.seed1 + self.seed2 + 33) % MAX_RAND_VALUE as u32;
+        self.seed1 = (self.seed1 * 3 + self.seed2) % MAX_RAND_VALUE;
+        self.seed2 = (self.seed1 + self.seed2 + 33) % MAX_RAND_VALUE;
         f64::from(self.seed1) / MAX_RAND_VALUE as f64
     }
 }
